@@ -3,9 +3,9 @@ package com.github.kongpf8848.viewworld.activity
 import android.os.Bundle
 import com.github.kongpf8848.viewworld.R
 import com.github.kongpf8848.viewworld.base.BaseActivity
-import kotlinx.android.synthetic.main.activity_arc_bg.*
+import com.github.kongpf8848.viewworld.databinding.ActivityArcBgBinding
 
-class ArcBgActivity : BaseActivity() {
+class ArcBgActivity : BaseActivity<ActivityArcBgBinding>() {
 
     override fun getLayoutId(): Int {
         return R.layout.activity_arc_bg
@@ -13,6 +13,6 @@ class ArcBgActivity : BaseActivity() {
 
     override fun onCreateEnd(savedInstanceState: Bundle?) {
         super.onCreateEnd(savedInstanceState)
-        toolbar.setNavigationOnClickListener { finish() }
+        binding.toolbar.setNavigationOnClickListener { finish() }
     }
 }
