@@ -4,12 +4,12 @@ import android.os.Bundle
 import android.os.Looper
 import com.github.kongpf8848.viewworld.R
 import com.github.kongpf8848.viewworld.base.BaseActivity
-import kotlinx.android.synthetic.main.activity_count_down.*
+import com.github.kongpf8848.viewworld.databinding.ActivityCountDownBinding
 
 /**
  * 倒计时文本
  */
-class CountDownActivity:BaseActivity(){
+class CountDownActivity : BaseActivity<ActivityCountDownBinding>() {
     override fun getLayoutId(): Int {
         return R.layout.activity_count_down
     }
@@ -17,19 +17,19 @@ class CountDownActivity:BaseActivity(){
     override fun onCreateEnd(savedInstanceState: Bundle?) {
         super.onCreateEnd(savedInstanceState)
 
-        tv_skip.callback=::onAnimationEnd
+        binding.tvSkip.callback = ::onAnimationEnd
         Looper.myQueue().addIdleHandler {
-            tv_skip.start()
+            binding.tvSkip.start()
             false
         }
     }
 
-    private fun onAnimationEnd(){
+    private fun onAnimationEnd() {
         finish()
     }
 
     override fun onStop() {
         super.onStop()
-        tv_skip.stop()
+        binding.tvSkip.stop()
     }
 }
