@@ -5,9 +5,9 @@ import android.widget.SeekBar
 import android.widget.SeekBar.OnSeekBarChangeListener
 import com.github.kongpf8848.viewworld.R
 import com.github.kongpf8848.viewworld.base.BaseActivity
-import kotlinx.android.synthetic.main.activity_font_setting.*
+import com.github.kongpf8848.viewworld.databinding.ActivityFontSettingBinding
 
-class FontSettingActivity: BaseActivity() {
+class FontSettingActivity : BaseActivity<ActivityFontSettingBinding>() {
 
     override fun getLayoutId(): Int {
         return R.layout.activity_font_setting
@@ -15,18 +15,18 @@ class FontSettingActivity: BaseActivity() {
 
     override fun onCreateEnd(savedInstanceState: Bundle?) {
         super.onCreateEnd(savedInstanceState)
-        toolbar.setNavigationOnClickListener {
+        binding.toolbar.setNavigationOnClickListener {
             finish()
         }
 
-        font_seekBar.setOnSeekBarChangeListener(object : OnSeekBarChangeListener {
+        binding.fontSeekBar.setOnSeekBarChangeListener(object : OnSeekBarChangeListener {
             override fun onProgressChanged(
                 seekBar: SeekBar,
                 value: Int,
                 bManual: Boolean
             ) {
-                val size= font_seekBar.getRawTextSize(value)
-                tv_message.textSize = size.toFloat()
+                val size = binding.fontSeekBar.getRawTextSize(value)
+                binding.tvMessage.textSize = size.toFloat()
             }
 
             override fun onStartTrackingTouch(seekBar: SeekBar) {}
