@@ -5,9 +5,9 @@ import com.github.kongpf8848.viewworld.R
 import com.github.kongpf8848.viewworld.activity.banner.BannerActivity
 import com.github.kongpf8848.viewworld.activity.slidemenu.SlideMenuActivity
 import com.github.kongpf8848.viewworld.base.BaseActivity
-import kotlinx.android.synthetic.main.activity_main.*
+import com.github.kongpf8848.viewworld.databinding.ActivityMainBinding
 
-class MainActivity : BaseActivity() {
+class MainActivity : BaseActivity<ActivityMainBinding>() {
 
     override fun getLayoutId(): Int {
         return R.layout.activity_main
@@ -15,35 +15,35 @@ class MainActivity : BaseActivity() {
 
     override fun onCreateEnd(savedInstanceState: Bundle?) {
         super.onCreateEnd(savedInstanceState)
-        button1.setOnClickListener {
+        binding.button1.setOnClickListener {
             startActivity(ArcBgActivity::class.java)
         }
-        button4.setOnClickListener {
+        binding.button4.setOnClickListener {
             startActivity(VerifyCodeActivity::class.java)
         }
-        button5.setOnClickListener {
+        binding.button5.setOnClickListener {
             startActivity(VerifyCode2Activity::class.java)
         }
-        button6.setOnClickListener {
+        binding.button6.setOnClickListener {
             startActivity(TabLayoutActivity::class.java)
         }
-        button7.setOnClickListener {
+        binding.button7.setOnClickListener {
             startActivity(BannerActivity::class.java)
         }
-        button8.setOnClickListener {
+        binding.button8.setOnClickListener {
             startActivity(ProgressActivity::class.java)
         }
-        button9.setOnClickListener {
+        binding.button9.setOnClickListener {
             startActivity(TextViewActivity::class.java)
         }
-        button9.setOnClickListener {
+        binding.button9.setOnClickListener {
             startActivity(TextViewActivity::class.java)
         }
-        button10.setOnClickListener {
+        binding.button10.setOnClickListener {
             startActivity(SlideMenuActivity::class.java)
         }
 
-        button11.setOnClickListener {
+        binding.button11.setOnClickListener {
             startActivity(SelectAvatarActivity::class.java)
         }
 
