@@ -3,19 +3,17 @@ package com.github.kongpf8848.viewworld.activity
 import android.os.Bundle
 import com.github.kongpf8848.viewworld.R
 import com.github.kongpf8848.viewworld.base.BaseActivity
-import com.kongpf.commonhelper.ScreenHelper
-import kotlinx.android.synthetic.main.activity_font_setting.*
-import kotlinx.android.synthetic.main.activity_font_setting.toolbar
-import kotlinx.android.synthetic.main.activity_textview.*
+import com.github.kongpf8848.viewworld.databinding.ActivityTextviewBinding
 
-class TextViewActivity : BaseActivity() {
+
+class TextViewActivity : BaseActivity<ActivityTextviewBinding>() {
     override fun getLayoutId(): Int {
         return R.layout.activity_textview
     }
 
     override fun onCreateEnd(savedInstanceState: Bundle?) {
         super.onCreateEnd(savedInstanceState)
-        toolbar.setNavigationOnClickListener {
+        binding.toolbar.setNavigationOnClickListener {
             finish()
         }
     }
