@@ -3,10 +3,10 @@ package com.github.kongpf8848.viewworld.activity.slidemenu
 import android.os.Bundle
 import com.github.kongpf8848.viewworld.R
 import com.github.kongpf8848.viewworld.base.BaseActivity
-import com.kongpf.commonhelper.ScreenHelper
-import kotlinx.android.synthetic.main.activity_slide_menu_1.*
+import com.github.kongpf8848.viewworld.databinding.ActivitySlideMenu1Binding
+import io.github.kongpf8848.commonhelper.ScreenHelper
 
-class SlideMenu1Activity: BaseActivity(){
+class SlideMenu1Activity : BaseActivity<ActivitySlideMenu1Binding>() {
 
     override fun getLayoutId(): Int {
         return R.layout.activity_slide_menu_1
@@ -14,9 +14,9 @@ class SlideMenu1Activity: BaseActivity(){
 
     override fun onCreateEnd(savedInstanceState: Bundle?) {
         super.onCreateEnd(savedInstanceState)
-        toolbar.setNavigationOnClickListener {
+        binding.toolbar.setNavigationOnClickListener {
             finish()
         }
-        slideMenu.setMenuWidth(ScreenHelper.dp2px(applicationContext,180f))
+        binding.slideMenu.setMenuWidth(ScreenHelper.dp2px(applicationContext, 180f))
     }
 }
