@@ -2,15 +2,10 @@ package com.github.kongpf8848.viewworld.activity.slidemenu
 
 import android.os.Bundle
 import com.github.kongpf8848.viewworld.R
-import com.github.kongpf8848.viewworld.activity.ArcBgActivity
-import com.github.kongpf8848.viewworld.activity.VerifyCodeActivity
 import com.github.kongpf8848.viewworld.base.BaseActivity
-import com.kongpf.commonhelper.ScreenHelper
-import kotlinx.android.synthetic.main.activity_main.*
-import kotlinx.android.synthetic.main.activity_slide_menu.*
-import kotlinx.android.synthetic.main.activity_slide_menu.button1
+import com.github.kongpf8848.viewworld.databinding.ActivitySlideMenuBinding
 
-class SlideMenuActivity: BaseActivity(){
+class SlideMenuActivity : BaseActivity<ActivitySlideMenuBinding>() {
 
     override fun getLayoutId(): Int {
         return R.layout.activity_slide_menu
@@ -18,13 +13,13 @@ class SlideMenuActivity: BaseActivity(){
 
     override fun onCreateEnd(savedInstanceState: Bundle?) {
         super.onCreateEnd(savedInstanceState)
-        toolbar.setNavigationOnClickListener {
+        binding.toolbar.setNavigationOnClickListener {
             finish()
         }
-        button1.setOnClickListener {
+        binding.button1.setOnClickListener {
             startActivity(SlideMenu1Activity::class.java)
         }
-        button2.setOnClickListener {
+        binding.button2.setOnClickListener {
             startActivity(SlideMenu2Activity::class.java)
         }
     }
