@@ -5,28 +5,36 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.annotation.Nullable
-import com.github.kongpf8848.viewworld.R
 import com.github.kongpf8848.viewworld.base.BaseFragment
-import kotlinx.android.synthetic.main.fragment_title.*
+import com.github.kongpf8848.viewworld.databinding.FragmentTitleBinding
 
 class TitleFragment : BaseFragment() {
 
-    companion object{
-        fun newInstance(title:String): TitleFragment {
+    private lateinit var binding: FragmentTitleBinding
+
+    companion object {
+        fun newInstance(title: String): TitleFragment {
             return TitleFragment().apply {
-                arguments=Bundle().apply {
-                    putString("title",title)
+                arguments = Bundle().apply {
+                    putString("title", title)
                 }
             }
         }
     }
-    override fun onCreateView(inflater: LayoutInflater, @Nullable container: ViewGroup?, @Nullable savedInstanceState: Bundle?): View? {
-       return inflater.inflate(R.layout.fragment_title,container,false)
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        @Nullable container: ViewGroup?,
+        @Nullable savedInstanceState: Bundle?
+    ): View? {
+        binding = FragmentTitleBinding.inflate(inflater, container, false)
+        val view = binding.root
+        return view
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        tv_subject.text = arguments?.getString("title", "")
+        binding.tvSubject.text = arguments?.getString("title", "")
     }
 
 }
