@@ -2,21 +2,19 @@ package com.github.kongpf8848.viewworld.activity
 
 import android.os.Bundle
 import android.os.Looper
-import android.text.TextUtils
-import androidx.core.widget.doAfterTextChanged
 import com.github.kongpf8848.viewworld.R
 import com.github.kongpf8848.viewworld.base.BaseActivity
+import com.github.kongpf8848.viewworld.databinding.ActivityVerifyCodeSquareBinding
 import com.github.kongpf8848.viewworld.utis.KeyboardUtils
 import com.github.kongpf8848.viewworld.utis.MainHandlerUtils
 import com.github.kongpf8848.viewworld.views.VerificationCodeEditText
-import com.kongpf.commonhelper.ToastHelper
-import kotlinx.android.synthetic.main.activity_font_setting.toolbar
-import kotlinx.android.synthetic.main.activity_verify_code_line.*
+import io.github.kongpf8848.commonhelper.ToastHelper
+
 
 /**
  * 验证码输入框
  */
-class VerifyCode2Activity : BaseActivity() {
+class VerifyCode2Activity : BaseActivity<ActivityVerifyCodeSquareBinding>() {
 
     override fun getLayoutId(): Int {
         return R.layout.activity_verify_code_square
@@ -24,9 +22,10 @@ class VerifyCode2Activity : BaseActivity() {
 
     override fun onCreateEnd(savedInstanceState: Bundle?) {
         super.onCreateEnd(savedInstanceState)
-        toolbar.setNavigationOnClickListener { finish() }
+        binding.toolbar.setNavigationOnClickListener { finish() }
 
-        et_verification_code.setOnInputTextListener(object: VerificationCodeEditText.OnInputTextListener{
+        binding.etVerificationCode.setOnInputTextListener(object :
+            VerificationCodeEditText.OnInputTextListener {
             override fun onInputTextComplete(text: CharSequence) {
                 ToastHelper.toastCenter("你输入的验证码为:${text}")
             }
@@ -34,7 +33,7 @@ class VerifyCode2Activity : BaseActivity() {
 
         Looper.myQueue().addIdleHandler {
             MainHandlerUtils.postDelayed({
-                KeyboardUtils.showSoftInput(et_verification_code)
+                KeyboardUtils.showSoftInput(binding.etVerificationCode)
             }, 500)
 
             false
